refactor: 게시글 수정/삭제 무효화 작업 및 quill 버전 변경
