+++ apps/next-client/package.json
@@ -31,6 +31,7 @@
     "react-icons": "^5.4.0",
     "react-markdown": "^9.0.3",
     "react-quill": "^2.0.0",
+    "react-quill-new": "^3.3.3",
     "react-toastify": "^11.0.2",
     "rehype-raw": "^7.0.0",
     "sanitize-html": "^2.14.0",
