+++ apps/next-client/src/components/community/PostActions.tsx
@@ -2,6 +2,7 @@
 
 import { useRouter } from 'next/navigation';
 import { useSession } from 'next-auth/react';
+import { useQueryClient } from '@tanstack/react-query';
 import { API_URLS } from '@/constants/urls';
 import { axiosInstance } from '@/services/axiosInstance';
 import { ALERT_MESSAGES } from '@/constants/alertMessage';
@@ -13,6 +14,7 @@ interface PostActionsProps {
 
 const PostActions = ({ postId, userId }: PostActionsProps) => {
   const router = useRouter();
+  const queryClient = useQueryClient();
   const { data: session, status } = useSession();
   const currentUserId = session?.user?.id;
 
@@ -28,6 +30,7 @@ const PostActions = ({ postId, userId }: PostActionsProps) => {
       });
 
       alert(ALERT_MESSAGES.SUCCESS.POST.POST_DELETE);
+      queryClient.invalidateQueries({ queryKey: ['posts'] });
       router.push('/community');
       
     } catch (error: unknown) {
